requirements.txt
autoflake==1.4
black==20.8b1
click==7.1.2
ffmpeg==1.4
flake8==3.8.4
isort==5.6.4
numba==0.50.1
numpy==1.19.0
pylint==2.6.0
pytest==6.2.2
pytest-cov==2.11.1
librosa==0.8.0
scipy==1.1.0
SoundFile==0.10.3.post1
torch==1.7.1
matplotlib==3.2.2
PyYAML==3.13
mypy==0.800
